+++ blocks/filter-select/src/edit.js
@@ -83,6 +83,8 @@ export default function Edit( object ) {
 					<div className="wp-personio-integration-selectcontrol-multiple">
 						{
 							<SelectControl
+                __next40pxDefaultSize
+                __nextHasNoMarginBottom
 								label={__('Choose filter', 'personio-integration-light')}
 								value={object.attributes.filter}
 								options={ personioTaxonomies }
@@ -92,16 +94,19 @@ export default function Edit( object ) {
 						}
 					</div>
 					<ToggleControl
+            __nextHasNoMarginBottom
 						label={__('Hide filter title', 'personio-integration-light')}
 						checked={ object.attributes.hideFilterTitle }
 						onChange={ value => onChangeHideFilterTitle( value, object ) }
 					/>
           <ToggleControl
+            __nextHasNoMarginBottom
             label={__('Hide submit button', 'personio-integration-light')}
             checked={ object.attributes.hideSubmitButton }
             onChange={ value => onChangeHideSubmitButton( value, object ) }
           />
           <ToggleControl
+            __nextHasNoMarginBottom
             label={__('Hide reset link', 'personio-integration-light')}
             checked={ object.attributes.hideResetLink }
             onChange={ value => onChangeHideResetLink( value, object ) }